==> priority/high_or_not_high/02_data_combining.ipynb <==
import pandas as pd

# Read CSV into a dataframe
high_priority = pd.read_csv("csv/final_high_priority_dataset.csv", index_col=0)
not_high_priority = pd.read_csv("csv/final_not_high_priority_dataset.csv", index_col=0)

high_priority

# Number or different labels
high_priority.labels.value_counts().to_frame()[:50]

# Number or different labels
not_high_priority.labels.value_counts().to_frame()[:50]

# Add a new column
high_priority["class"] = "high_priority"
high_priority

not_high_priority["class"] = "not_high_priority"
not_high_priority

all_data = pd.concat([high_priority, not_high_priority] , ignore_index= True)

all_data

all_data_no_dupes = all_data.drop_duplicates()
all_data_no_dupes

# Convert dataframe to a CSV file.
all_data_no_dupes.to_csv("csv/full_dataset_priority_high_or_not_high.csv")
