==> priority/high_or_not_high/05_priority_and_TD.ipynb <==
import pandas as pd
from transformers import pipeline, AutoTokenizer

# Load the Pre-trained Model and its Tokenizer
model_path = './model_local'
tokenizer = AutoTokenizer.from_pretrained(model_path)
classifier = pipeline("text-classification", model=model_path, tokenizer=tokenizer, truncation=True)

classifier("after user sing up the email must be confirmed prepare a goodlooking email template develop a be to send this template after user registration develop an endpoint to check confirmation token develop a fe page with confirmation status create a cron to check accounts which havent confirmed an email and remove them")

classifier("description of the problem when i try to perform file upload in sut via the browsers running in docker the file upload popup is not opening and neither accepts any files causing scripts to fail browser and version safari operating system ran on docker webdrivermanager version v503")

classifier("descriptive summary hyrax 302 makes rails 526 support official and rails 526 includes a security update for polymorphicpath which is a method used in all link generation across rails linkto urlfor etc all these methods no longer support string parameters which must be converted to symbols this is a problem in 526 and 5246 our current 525 is unaffected and upgrading may see problems on random pages we should check through the application for calls to polymorphicpath urlfor linkto and any other link generating methods you can think of all hyrax links should be fixed in 302 then we can safely update to hyrax 302 and rails 526 expected behavior all links on site call their generation method with symbols or objects not strings hyrax 302 rails 526 related work blocked 1578 from going to 302 accessibility concerns")

df = pd.read_csv("csv/TD_dataset_clean.csv" , index_col = 0)
df.info()

df['text'] = df['text'].astype(str)
type(df['text'].iloc[0])

df = df[df['label']==1]
df = df.reset_index(drop=True)
df

df.rename(columns={'label': 'td_label'}, inplace=True)


df_small = df.iloc[:5000]
df_small

# Function to get predictions and scores
def get_predictions(text):
    result = classifier(text)
    return result[0]['label'], result[0]['score']

# Apply the classifier to each row
df[['high_priority_label', 'high_priority_score']] = df.apply(lambda row: pd.Series(get_predictions(row['text'])), axis=1)
df

# Function to assign the new name based on high_priority_label
def assign_priority_name(label):
    if label == 'LABEL_0':
        return 'High Priority'
    elif label == 'LABEL_1':
        return 'Not High Priority'
    else:
        return 'Unknown'  # You can change this to handle other unexpected labels

# Apply the function to each row
df['priority_name'] = df['high_priority_label'].apply(assign_priority_name)
df

average_scores = df.groupby('priority_name')['high_priority_score'].mean()
average_scores
